Rename plotting v2 to just plotting
