+++ src/feedback_forensics/app/plotting/__init__.py
@@ -1,0 +1,1 @@
+from feedback_forensics.app.plotting.main import generate_plot
