+++ src/feedback_forensics/app/plotting/table.py
