--- src/feedback_forensics/app/plotting_v2/__init__.py
@@ -1,1 +1,0 @@
-from feedback_forensics.app.plotting_v2.main import generate_plot
